MNT: rename submodule
